wip track home venue playlist + more generic code
